[rtloader] Use upstream workaround for setting Python bin path on Windows
Due to an [upstream bug](https://bugs.python.org/issue34725) in how
Windows handles `Py_SetProgramName`, we have to use the suggested
workaround and also use `_Py_SetProgramFullPath` for now to make this
code work on that platform.

Since this is a workaround API in CPython, it is depricated as of
Python 3.11 and will be removed at some later version. Before it gets
removed, we should convert our Python 3 initialization code to the
more-appropriate `PyConfig API`
(https://docs.python.org/3.11/c-api/init_config.html#c.PyConfig).